mm: add IOMMU context to mm_struct
When binding an mm to a device context, the IOMMU driver needs to check if
a PASID (Process Address Space ID) has already been allocated for that mm
and, if there hasn't been any, allocate a new one. Rather than iterating
over all PASIDs (that is, visiting 1 << 20 IDR nodes in the worst case),
store a pointer to the IOMMU context inside mm_struct. The pointer is
protected by a global lock in the IOMMU subsystem, to handle multiple
threads binding the same mm to different devices concurrently.

Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>