uacce: Remove mm_exit() op
The mm_exit() op will be removed from the SVA API. When a process dies
and its mm goes away, the IOMMU driver won't notify us. We just need to
be prepared to handle a lot more aborted DMA. On the upside, it does
greatly simplify the queue management.

The uacce_mm struct, that tracks all queues bound to an mm, was only
used by the mm_exit() callback. Remove it.

Note that we also fix a leak in uacce_remove.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>